reorganize col_attribute.cpp
